// backend/src/dtos/user/health/health-details.request.dto.ts
import {
  ActivityLevel,
  DietType,
  GoalType,
  TimelineType,
} from "../../../types/health.types";

export interface HealthDetailsRequestDto {
  heightCm: number;
  weightKg: number;
  activityLevel: ActivityLevel;
  dietType: DietType;
  goal: GoalType;
  targetWeightKg?: number;
  preferredTimeline: TimelineType;

  dailyWaterIntakeLiters: number;
  sleepDurationHours: number;
}